+++ .vscode/settings.json
@@ -1,5 +1,3 @@
 {
-  "cSpell.words": [
-    "Populator"
-  ]
+  "cSpell.words": ["Populator"]
 }
